+++ lib/secretary.dart
@@ -4,6 +4,7 @@ export 'src/execution_params.dart';
 export 'src/recurring_task.dart';
 export 'src/result.dart';
 export 'src/secretary.dart';
+export 'src/state.dart';
 export 'src/task.dart';
 export 'src/task_overrides.dart';
 export 'src/types.dart';